requirements.txt
langchain>=0.3.19,<0.4.0
langchain-community>=0.3.18,<0.4.0
faiss-cpu>=1.10.0,<2.0.0
pymupdf>=1.25.3,<2.0.0
pypdf>=5.3.0,<6.0.0
sentence-transformers>=3.4.1,<4.0.0
rich>=13.9.4,<14.0.0
wget>=3.2,<4.0
python-dotenv>=1.0.1,<2.0.0
cryptography>=44.0.1,<45.0.0
ollama>=0.4.7,<0.5.0
ipykernel>=6.29.5,<7.0.0
langchain-huggingface>=0.1.2,<0.2.0
langchain-unstructured>=0.1.6,<0.2.0
unstructured>=0.16.23,<0.17.0
langchain-ollama>=0.2.3,<0.3.0
dataparser>=0.0.2,<0.0.3
dateparser>=1.2.1,<2.0.0
tavily-python>=0.7.1,<0.8.0
matplotlib>=3.10.1,<4.0.0
